requirements.txt
gradio
pytesseract
pdf2image
pillow
spacy
pandas